Linear model to predict holocellulose content from MIR spectra from Hodgkins et al. (2018)

irp_content_h_hodgkins_model is a linear model to predict holocellulose mass fractions [g/g] in samples based on normalized and area corrected peak heights of peaks identified in spectra. This model is meant for internal use in irp_content_klh_hodgkins(). The model was originally developed by Hodgkins et al. (2018) and is here reproduced from the same data (available via ir::ir_sample_data), but handling mass fraction in g/g instead of as percentages. Note that the models have been shown to be biased for peat samples (Teickner and Knorr 2022) .

Usage

irp_content_h_hodgkins_model

Format

An object of class stats::lm().

Source

The data set was derived from https://www.nature.com/articles/s41467-018-06050-2 and published by Hodgkins et al. (2018) under the CC BY 4.0 license https://creativecommons.org/licenses/by/4.0/. Hodgkins et al. (2018) originally derived the data on holocellulose content from De La Cruz, Florentino B. et al. (2016) https://www.liebertpub.com/doi/full/10.1089/ees.2014.0402. The linear model is reproduced in slightly modified form (see description) from Hodgkins et al. (2018) .

Note

Note that the models have been shown to be biased for peat samples (Teickner and Knorr 2022) .

References

De La Cruz, Florentino B., Osborne J, Barlaz MA (2016). “Determination of Sources of Organic Matter in Solid Waste by Analysis of Phenolic Copper Oxide Oxidation Products of Lignin.” Journal of Environmental Engineering, 142(2), 04015076. ISSN 0733-9372, doi:10.1061/(ASCE)EE.1943-7870.0001038 .

Hodgkins SB, Richardson CJ, Dommain R, Wang H, Glaser PH, Verbeke B, Winkler BR, Cobb AR, Rich VI, Missilmani M, Flanagan N, Ho M, Hoyt AM, Harvey CF, Vining SR, Hough MA, Moore TR, Richard PJH, De La Cruz, Florentino B., Toufaily J, Hamdan R, Cooper WT, Chanton JP (2018). “Tropical peatland carbon storage linked to global latitudinal trends in peat recalcitrance.” Nature communications, 9(1), 3640. doi:10.1038/s41467-018-06050-2 .

Teickner H, Knorr K (2022). “Improving Models to Predict Holocellulose and Klason Lignin Contents for Peat Soil Organic Matter with Mid Infrared Spectra.” Soil and methods. doi:10.5194/soil-2022-27 .