Configuration files defining preprocessing parameters.

The different configuration files are used in the models with the respective name. See irp_preprocess().

Usage

model_klason_lignin_2_config

model_holocellulose_2_config

model_eac_1_config

model_edc_1_config

Format

A list with the following elements:

irp_preprocess

A list with the a subset of the following elements:

do_interpolate

See irp_preprocess().

interpolate_start

See irp_preprocess().

interpolate_dw

See irp_preprocess().

do_clip

See irp_preprocess().

clip_range

See irp_preprocess().

do_interpolate_region

See irp_preprocess().

interpolate_region_range

See irp_preprocess().

do_bc

See irp_preprocess().

bc_method

See irp_preprocess().

bc_degree

See irp_preprocess().

bc_cutoff

See irp_preprocess().

do_smooth

See irp_preprocess().

smooth_method

See irp_preprocess().

smooth_p

See irp_preprocess().

smooth_n

See irp_preprocess().

smooth_m

See irp_preprocess().

smooth_ts

See irp_preprocess().

smooth_k

See irp_preprocess().

do_normalise

See irp_preprocess().

normalise_method

See irp_preprocess().

do_bin

See irp_preprocess().

bin_width

See irp_preprocess().

do_scale

See irp_preprocess().

scale_center

See irp_preprocess(). This is used to preprocess the training data before model computation. The values in data_scale are used to scale data used for prediction.

scale_scale

See irp_preprocess(). This is used to preprocess the training data before model computation. The values in data_scale are used to scale data used for prediction.

data_scale

Values used for scaling data used for prediction. A list with the following elements:

y_center

A numeric value representing the value to subtract from the dependent variable of a model during scaling. Corresponds to argument center in base::scale().

y_scale

A numeric value representing the value by which the dependent variable of a model is divided during scaling. Corresponds to argument scale in base::scale().

x_center

The same as y_center, but for the independent variable(s). Can be a numeric vector with an element for each independent variable.

y_scale

The same as y_center, but for the independent variable(s). Can be a numeric vector with an element for each independent variable.

An object of class list of length 2.

An object of class list of length 2.

An object of class list of length 2.